+++ frontend/src/components/Place/PlaceCard.jsx
